aggregator/requirements.txt
aio-pika==9.5.7
httpx==0.28.1
planetary-computer==1.0.0
pydantic==2.11.7
python-dotenv==1.1.1